requirements.txt
Flask
numpy
tensorflow
Pillow
google-cloud-firestore
